cr lf optimiert damit ack in gleicher zeile wie telegramm ist (mit .hex)
git-svn-id: file:///usr/local/svn/freebus/trunk@1589 219b8a4f-0e80-dc11-b224-0019d16aef49

+++ 89LPC922/busmon/fb_app_mon.c
@@ -34,7 +34,7 @@ void read_value_req(void)	// entfällt
 {
 }
 
-unsigned int read_obj_value(unsigned char objno)	// entfällt
+unsigned long read_obj_value(unsigned char objno)	// entfällt
 {
 	objno;
 	return(0);
