+++ 89LPC922/busmon/fb_app_mon.h
@@ -19,7 +19,7 @@
 void restart_app(void);		// Alle Applikations-Parameter zurücksetzen
 void write_value_req(void);	// entfällt
 void read_value_req(void);	// entfällt
-unsigned int read_obj_value(unsigned char objno);	// entfällt
+unsigned long read_obj_value(unsigned char objno);	// entfällt
 void write_obj_value(unsigned char objno, unsigned int val);			// entfällt
 
 #endif
